refactor upgrade function
