+++ app/app.go
@@ -724,7 +724,7 @@ func initParamsKeeper(
 
 func (app *App) applyUpgrades() {
 	app.applyUpgrade_v0_1_2()
-	app.applyUpgrade_v0_2_2()
+	app.applyUpgrade_v0_2_4()
 	// app.applyUpgrade_v0_2_3()
 }
