+++ app/upgrade_v0_2_2.go
@@ -4,20 +4,20 @@ import (
 	"context"
 
 	upgradetypes "cosmossdk.io/x/upgrade/types"
-	"github.com/Galactica-corp/galactica/app/upgrades/v0_2_2"
+	"github.com/Galactica-corp/galactica/app/upgrades/v0_2_4"
 
 	tmproto "github.com/cometbft/cometbft/proto/tendermint/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/module"
 )
 
-func (app *App) applyUpgrade_v0_2_2() {
-	planName := v0_2_2.UpgradeName
+func (app *App) applyUpgrade_v0_2_4() {
+	planName := v0_2_4.UpgradeName
 
 	logger := app.Logger().With("upgrade", planName)
 
-	app.UpgradeKeeper.SetUpgradeHandler(planName, app.upgradeHandler_v0_2_2())
-	
+	app.UpgradeKeeper.SetUpgradeHandler(planName, app.upgradeHandler_v0_2_4())
+
 	ctx := app.NewContext(true).WithBlockHeader(tmproto.Header{Height: app.LastBlockHeight()}).
 		WithMultiStore(app.CommitMultiStore())
 
@@ -28,18 +28,18 @@ func (app *App) applyUpgrade_v0_2_2() {
 	}
 
 	if doneHeight != 0 {
-		logger.Info("upgrade v0.2.2 done")
+		logger.Info("upgrade v0.2.4 done")
 		return
 	}
 
 	logger.Info("Schedule upgrade plan", "name", planName)
 
-	if err := app.UpgradeKeeper.ScheduleUpgrade(ctx, v0_2_2.Plan); err != nil {
+	if err := app.UpgradeKeeper.ScheduleUpgrade(ctx, v0_2_4.Plan); err != nil {
 		panic(err)
 	}
 }
 
-func (app *App) upgradeHandler_v0_2_2() func(ctx context.Context, _ upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
+func (app *App) upgradeHandler_v0_2_4() func(ctx context.Context, _ upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
 	return func(ctx context.Context, plan upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
 		logger := sdk.UnwrapSDKContext(ctx).Logger()
