+++ app/upgrades/v0_2_4/upgrade.go
@@ -1,11 +1,11 @@
-package v0_2_2
+package v0_2_4
 
 import (
 	upgradetypes "cosmossdk.io/x/upgrade/types"
 )
 
 const (
-	UpgradeName        = "0.2.2"
+	UpgradeName        = "0.2.4"
 	UpgradeBlockHeight = 1
 )
